Merge duplicate bugs in recheckwatch
When a bug is marked as a duplicate move all the hits to the
duplicate_of bug and delete the duplicate bug.

This change should make it easier to keep recheckwatch easier to read
by removing duplicate bugs.  The fewer bugs on recheck the easier it is
for a patch author to figure out which one fits his issue.

Note: unmarking a bug as a duplicate will not undo this

Change-Id: I19b25bb9595330393e664e9499204083b18ab5c9