Delete closed bugs after 5 days in recheckwatch
Instead of leaving closed bugs around for 30 days, remove them after 5
days.  If a closed bug hasn't reoccurred in 5 days there is a good chance
it has been fixed correctly.  The number of days to keep around a closed
bug is set in by closed_age in the config file.

Change-Id: I5d136173ca14a8f9e77a44e29e9499c94a417cd3